Made the rocket simulation more modular for later documentatino
