+++ backend/simulation/rocket_sim/__init__.py
@@ -1,4 +1,5 @@
 from .rocket import create_rocket, create_hybrid_motor
+from .engine import create_hybrid_motor
 from .flight import run_flight
 from .environment import create_environment
 from .fuel import create_oxidizer_tank
